Cleaned physics implementation
- Created apply force function
- Added doxygen comments
- Vector2 now uses floats instead of double

+++ Engine/src/Component/ShapeRenderer.cpp
@@ -7,7 +7,7 @@
 
 namespace
 {
-constexpr double kEpsilon = 1e-6;
+constexpr float kEpsilon = 1e-6;
 }
 
 ShapeRenderer& ShapeRenderer::setColor(const Color& newColor)
@@ -16,7 +16,7 @@ ShapeRenderer& ShapeRenderer::setColor(const Color& newColor)
 	return *this;
 }
 
-ShapeRenderer& ShapeRenderer::setCircle(double newRadius)
+ShapeRenderer& ShapeRenderer::setCircle(float newRadius)
 {
 	type = ShapeRenderType::Circle;
 	radius = std::max(newRadius, kEpsilon);
@@ -36,7 +36,7 @@ Color ShapeRenderer::getColor() const
 	return color;
 }
 
-double ShapeRenderer::getRadius() const
+float ShapeRenderer::getRadius() const
 {
 	return radius;
 }
